+++ test/base/babel-tests.js
@@ -1072,3 +1072,25 @@ test("actions are not reassignable", () => {
         a.m4 = () => {}
     }).toThrow(/Cannot assign to read only property 'm4'/)
 })
+
+test("it should support asyncAction as decorator (babel)", async () => {
+    const values = []
+
+    mobx.configure({ enforceActions: true })
+
+    class X {
+        @observable a = 1
+
+        @mobx.flow
+        f = function*(initial) {
+            this.a = initial // this runs in action
+            this.a += yield Promise.resolve(5)
+            this.a = this.a * 2
+            return Promise.resolve(this.a)
+        }
+    }
+
+    const x = new X()
+
+    expect(await x.f(3)).toBe(16)
+})